Добавил глобальный баннер ошибки (не везде работает). Миграцию поправил. Баги поправил если нет авторизации.

+++ database/migrations/2024_06_11_074809_create_animes_table.php
@@ -34,7 +34,7 @@ return new class extends Migration {
             $table->string('author')->nullable();
             $table->integer('shiki_id')->nullable();
             $table->integer('mal_id')->nullable();
-            $table->integer('shiki_score')->nullable();
+            $table->float('shiki_score')->nullable();
             $table->timestamps();
         });
     }
